+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+#### independent (2021-07-21)
+
+##### Refactors
+
+-   **@angular-ru/common:** changed name on inputFilter (#780) (e572d59e)
+
 #### independent (2021-07-20)
 
 ##### New Features